+++ tests/testthat/test_plots.R
@@ -4,28 +4,22 @@ context("Plot results")
 data("cmapPerturbationsCompounds")
 perturbations <- cmapPerturbationsCompounds
 data("diffExprStat")
-cellLine <- "HEPG2"
 
-compareKnockdown <- list()
-compareKnockdown$pearson  <- compareAgainstCMap(diffExprStat, perturbations,
-                                                cellLine, method="pearson")
-compareKnockdown$spearman <- compareAgainstCMap(diffExprStat, perturbations,
-                                                cellLine, method="spearman")
-compareKnockdown$gsea     <- compareAgainstCMap(diffExprStat, perturbations,
-                                                cellLine, method="gsea")
-condition <- compareKnockdown$gsea$compounds[[1]]
+compareKnockdown <- compareAgainstCMap(diffExprStat, perturbations)
+condition <- compareKnockdown[1, 1]
 
-test_that("Plot Spearman correlation", {
-    plot <- plotCMapComparison(compareKnockdown$spearman, condition)
+test_that("Plot Spearman's correlation", {
+    plot <- plotCMapComparison(compareKnockdown, condition, "spearman")
     expect_is(plot, "NULL")
 })
 
-test_that("Plot Pearson correlation", {
-    plot <- plotCMapComparison(compareKnockdown$pearson, condition)
+test_that("Plot Pearson's correlation", {
+    plot <- plotCMapComparison(compareKnockdow, condition, "pearson")
     expect_is(plot, "NULL")
 })
 
 test_that("Plot GSEA", {
-    plot <- plotCMapComparison(compareKnockdown$gsea, condition)
+    plot <- plotCMapComparison(compareKnockdown, condition, "gsea")
     expect_is(plot, "ggplot")
 })
+