+++ NEWS.md
@@ -30,7 +30,7 @@ function names:
     - By default, calculate mean across cell lines if there is more than one 
     cell line available; disabled if argument `cellLineMean = FALSE`
     - Allow to rank (or not) individual cell line perturbations (argument
-    `rankIndividualCellLines`) if mean is calculated
+    `rankIndividualCellLinePerturbations`) when the mean is calculated
     - Allow to perform multiple comparison methods if desired (by providing a 
     vector of supported methods via the `method` argument)
 * Improve plotting of CMap comparisons (`plot()`):