+++ NEWS.md
@@ -1,4 +1,4 @@
-# 1.0.3 (2 May, 2019)
+# 1.2.1 (22 May, 2019)
 
 * Update the tutorial and function documentation
 * Replace all references to `L1000` with `CMap`, as appropriate, including in
@@ -73,6 +73,10 @@ function names:
     according to new internal changes and fix their respective code in the 
     documentation
 * Include license and copyright text for cmapR code
+
+# 1.0.3 (3 December, 2018)
+
+* Add tag ImmunoOncology to BiocViews
 
 # 1.0.2 (11 November, 2018)